ml_service/requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
opencv-python==4.9.0.80
mediapipe==0.10.9
numpy==1.26.4
scipy==1.12.0
python-multipart==0.0.9
pillow==10.2.0